Tests/LibWeb/Text/expected/XHR/XMLHttpRequest-synchronous-send-from-microtask.txt

data: status=200 responseText=hello
http: status=200 responseText=hello over http
